docker-flask/requirements.txt
argon2-cffi==21.1.0
cachelib==0.4.1
certifi==2021.10.8       
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==35.0.0
Flask==2.0.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn
idna==3.3
itsdangerous
Jinja2==3.0.2
jsonify
MarkupSafe==2.0.1
mysqlclient
numba
numpy
pandas
pycparser
PyJWT
python-binance
python-dateutil
pytz
requests
SQLAlchemy
twilio
urllib3
Werkzeug